requirements.txt
aiosmtpd==1.4.2
asgiref==3.4.1
atpublic==2.3
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
dj-database-url==0.5.0
Django==3.2.10
django-braces==1.15.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
fontawesomefree==6.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
psycopg2==2.9.3
PyJWT==1.7.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
